refactor: abstract asset modal logic into different hook
