+++ src/components/common/Swap/components/CoinListItem/CoinListItem.tsx
@@ -8,6 +8,7 @@ import {Tooltip} from "react-tooltip";
 import "react-tooltip/dist/react-tooltip.css";
 import styles from "./CoinListItem.module.css";
 import {useAssetImage} from "@/src/hooks/useAssetImage";
+import Image from "next/image";
 
 type Props = {
   assetData: Omit<CoinDataWithPrice, "price"> & {
@@ -24,9 +25,13 @@ const CoinListItem = ({assetData}: Props) => {
   return (
     <span className={clsx(styles.coin, !assetData?.name && styles.centered)}>
       <Tooltip id="verified-tooltip" />
-      <img
+
+      <Image
         src={assetData.icon || fallbackIcon}
         alt={`${assetData.name} icon`}
+        width={40}
+        height={40}
+        priority
       />
 
       <div className={styles.names}>
